// frontend/src/app/core/generated/circabc/model/bulkInviteData.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NameValue } from './nameValue';
import { BulkImportUserData } from './bulkImportUserData';


export interface BulkInviteData { 
    bulkImportUserData: Array<BulkImportUserData>;
    igProfiles: Array<NameValue>;
}