// frontend/src/app/core/generated/circabc/model/interestGroupProfile.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InterestGroup } from './interestGroup';
import { Profile } from './profile';


/**
 * the representation of the membership of a user with an Interest group
 */
export interface InterestGroupProfile { 
    profile?: Profile;
    interestGroup?: InterestGroup;
}